requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
exceptiongroup==1.3.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uvloop==0.21.0
yarl==1.20.1
